requirements.txt
asgiref==3.7.2
Django==5.0.3
psycopg2-binary==2.9.9
sqlparse==0.4.4
django-allauth
djangorestframework
markdown
django-filter
dj_rest_auth
djangorestframework-jwt